requirements.txt
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.3
idna==3.2
Pillow==8.3.1
requests==2.26.0
soupsieve==2.2.1
urllib3==1.26.6
